finished refactoring of PluginFramework exceptions.
