+++ CLClient/src/main/java/ws/argo/CLClient/ArgoClientContext.java
@@ -19,7 +19,7 @@ import javax.ws.rs.client.WebTarget;
 import net.dharwin.common.tools.cli.api.CLIContext;
 import net.dharwin.common.tools.cli.api.console.Console;
 import ws.argo.CLClient.config.ClientConfiguration;
-import ws.argo.plugin.transport.sender.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
 import ws.argo.probe.Probe;
 
 /**
