+++ ProbeSender/pom.xml
@@ -17,11 +17,6 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>ws.argo.wireline</groupId>
-			<artifactId>ArgoWirelineFormat</artifactId>
-			<version>0.4.1-SNAPSHOT</version>
-		</dependency>
-		<dependency>
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
 			<version>2.4</version>
