+++ ProbeSender/src/main/java/ws/argo/transport/probe/standard/MulticastTransport.java
@@ -15,8 +15,8 @@ import java.util.logging.Logger;
 import javax.xml.bind.JAXBException;
 
 import ws.argo.plugin.transport.sender.Transport;
-import ws.argo.plugin.transport.sender.TransportConfigException;
-import ws.argo.plugin.transport.sender.TransportException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportException;
 import ws.argo.probe.Probe;
 import ws.argo.probe.ProbeSenderException;
