+++ ProbeSender/src/test/java/ws/argo/probe/test/ProbeSenderTest.java
@@ -14,8 +14,8 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import ws.argo.plugin.transport.sender.TransportConfigException;
-import ws.argo.plugin.transport.sender.TransportException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportException;
 import ws.argo.probe.Probe;
 import ws.argo.probe.Probe;
 import ws.argo.probe.ProbeSender;
