+++ Responder/ResponderDaemon/src/main/java/ws/argo/responder/Responder.java
@@ -50,7 +50,7 @@ import ws.argo.plugin.probehandler.ProbeHandlerConfigException;
 import ws.argo.plugin.probehandler.ProbeHandlerPluginIntf;
 import ws.argo.plugin.transport.responder.ProbeProcessor;
 import ws.argo.plugin.transport.responder.Transport;
-import ws.argo.plugin.transport.responder.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
 import ws.argo.wireline.probe.ProbeWrapper;
 
 /**
