+++ Responder/ResponderDaemon/src/main/java/ws/argo/responder/transport/sns/SNSListenerResource.java
@@ -29,7 +29,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
-import ws.argo.plugin.transport.responder.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
 import ws.argo.responder.transport.AmazonSNSTransport;
 import ws.argo.wireline.probe.ProbeParseException;
 import ws.argo.wireline.probe.ProbeWrapper;
