+++ Responder/ResponderDaemon/src/test/java/ws/argo/responder/test/ResponderProbeTest.java
@@ -10,8 +10,8 @@ import org.glassfish.grizzly.http.server.HttpServer;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 
-import ws.argo.plugin.transport.sender.TransportConfigException;
-import ws.argo.plugin.transport.sender.TransportException;
+import ws.argo.plugin.transport.exception.TransportConfigException;
+import ws.argo.plugin.transport.exception.TransportException;
 import ws.argo.probe.ProbeSender;
 import ws.argo.probe.ProbeSenderException;
 import ws.argo.probe.ProbeSenderFactory;
